requirements.txt
Django==5.2.6
django-webpush==0.3.6
pywebpush==2.1.0
webpush==1.0.5
Pillow==10.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-decouple==3.8
whitenoise==6.6.0
